requirements.txt
discord.py>=2.0.0
databases>=0.5.0
db-sqlite3>=0.0.1
